+++ changelog.md
@@ -1,1 +1,1 @@
-- Added Brazilian Portuguese translation
+- Added recipe advancements