+++ changelog.md
@@ -1,2 +1,1 @@
-- Added message to notify the user when the dimension hasn't been created yet
-- Used ITeleporter implementation
+- Updated to 1.16.4